effects: extract from items.c
