+++ src/game/effects.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "global/types.h"
+
+void Effect_InitialiseArray(void);
+int16_t Effect_Create(int16_t room_num);
+void Effect_Kill(int16_t fx_num);
+void Effect_NewRoom(int16_t fx_num, int16_t room_num);
