+++ src/game/effects/gun.c
@@ -2,7 +2,7 @@
 
 #include "3dsystem/phd_math.h"
 #include "game/draw.h"
-#include "game/items.h"
+#include "game/effects.h"
 #include "game/random.h"
 #include "global/vars.h"
