+++ src/game/effects/gunshot.c
@@ -1,7 +1,7 @@
 #include "game/objects/effects/gunshot.h"
 
+#include "game/effects.h"
 #include "game/effects/blood.h"
-#include "game/items.h"
 #include "game/objects/effects/ricochet.h"
 #include "game/random.h"
 #include "game/sound.h"
