+++ src/game/objects/ai/natla.c
@@ -3,8 +3,8 @@
 #include "3dsystem/phd_math.h"
 #include "game/box.h"
 #include "game/collide.h"
+#include "game/effects.h"
 #include "game/effects/gun.h"
-#include "game/items.h"
 #include "game/lot.h"
 #include "game/music.h"
 #include "game/people.h"
