+++ src/game/objects/effects/body_part.c
@@ -2,6 +2,7 @@
 
 #include "3dsystem/phd_math.h"
 #include "game/control.h"
+#include "game/effects.h"
 #include "game/items.h"
 #include "game/lara.h"
 #include "game/room.h"
