+++ src/game/objects/effects/ricochet.c
@@ -1,6 +1,6 @@
 #include "game/objects/effects/ricochet.h"
 
-#include "game/items.h"
+#include "game/effects.h"
 #include "game/random.h"
 #include "game/sound.h"
 #include "global/vars.h"
