+++ src/game/objects/traps/dart.c
@@ -3,6 +3,7 @@
 #include "game/collide.h"
 #include "game/control.h"
 #include "game/draw.h"
+#include "game/effects.h"
 #include "game/effects/blood.h"
 #include "game/items.h"
 #include "game/random.h"
