+++ src/game/setup.c
@@ -4,11 +4,11 @@
 #include "config.h"
 #include "game/cinema.h"
 #include "game/draw.h"
+#include "game/effects.h"
 #include "game/gamebuf.h"
 #include "game/gameflow.h"
 #include "game/hair.h"
 #include "game/inv.h"
-#include "game/items.h"
 #include "game/lara.h"
 #include "game/level.h"
 #include "game/lot.h"
